requirements.txt
pyyaml
streamlit
emoji
pandas
joblib
sklearn